requirements.txt
opencv-python
mediapipe
pyautogui
numpy
pyyaml
pystray
pillow
keyboard
pyyaml
